requirements.txt
# AWS SAA-C03 30일 커리큘럼 시스템 의존성

# 데이터 검증
pydantic>=2.0.0

# 템플릿 엔진
jinja2>=3.1.0

# 설정 파일 처리
pyyaml>=6.0

# 테스트
pytest>=7.0.0
pytest-cov>=4.0.0

# AWS SDK (테스트용)
boto3>=1.26.0

# AWS 서비스 모킹
moto>=4.0.0

# 타입 체킹 (개발용)
mypy>=1.0.0

# 코드 포맷팅 (개발용)
black>=23.0.0
isort>=5.12.0

# 린팅 (개발용)
flake8>=6.0.0
pylint>=2.17.0
